Add empty checks and type checks for new type safety constraints
Fix docblock annotations to match parameter types

Explicitly declare previously ‘dynamic’ class properties

+++ security/Password.php
@@ -196,8 +196,8 @@ class Password {
 		} while (1);
 
 		$result = '$2a$';
-		$result .= chr(ord('0') + $count / static::BF);
-		$result .= chr(ord('0') + $count % static::BF);
+		$result .= chr((int) (ord('0') + $count / static::BF));
+		$result .= chr((int) (ord('0') + $count % static::BF));
 		$result .= '$' . $output;
 
 		return $result;
